backend/cars/urls.py

from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import (
    BookingViewSet,
    CarViewSet,
    login_view,
    logout_view,
    me_view,
    signup_view,
)

router = DefaultRouter()
router.register(r'cars', CarViewSet)
router.register(r'bookings', BookingViewSet)

urlpatterns = [
    path("", include(router.urls)),
    path("auth/signup/", signup_view),
    path("auth/login/", login_view),
    path("auth/logout/", logout_view),
    path("auth/me/", me_view),
]
